Fixes issue #2306 - Rename standard/slim extensions to better reflect what they do
