+++ extension/micro/src/main/java/module-info.java
@@ -32,11 +32,11 @@ import cloud.piranha.core.api.WebApplicationExtension;
 module cloud.piranha.extension.micro {
     provides WebApplicationExtension with MicroExtension;
     requires cloud.piranha.extension.apache.fileupload;
-    requires cloud.piranha.extension.async;
     requires cloud.piranha.extension.herring;
     requires cloud.piranha.extension.policy;
     requires cloud.piranha.extension.security.jakarta;
     requires cloud.piranha.extension.security.servlet;
+    requires cloud.piranha.extension.standard.async;
     requires cloud.piranha.extension.standard.localeencoding;
     requires cloud.piranha.extension.standard.mimetype;
     requires cloud.piranha.extension.standard.servletannotations;
