+++ extension/standard-async/pom.xml
@@ -2,18 +2,14 @@
 
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-
     <parent>
         <groupId>cloud.piranha.extension</groupId>
         <artifactId>project</artifactId>
         <version>22.1.0-SNAPSHOT</version>
     </parent>
-
-    <artifactId>piranha-extension-async</artifactId>
+    <artifactId>piranha-extension-standard-async</artifactId>
     <packaging>jar</packaging>
-
-    <name>Piranha - Extension - Async</name>
-
+    <name>Piranha - Extension - Standard Async</name>
     <dependencies>
         <!-- compile -->
         <dependency>
@@ -26,6 +22,11 @@
         <dependency>
             <groupId>org.junit.jupiter</groupId>
             <artifactId>junit-jupiter-api</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.junit.platform</groupId>
+            <artifactId>junit-platform-launcher</artifactId>
             <scope>test</scope>
         </dependency>
     </dependencies>
